requirements-dev.txt
coverage
django-coverage
django-debug-toolbar
-r requirements.txt
